Ordering (#260)
This PR adds ordering capability in the API and web UI. There is also some refactoring of the web UI that has significantly improved performance.

To get proper ordering (particularly by seeders/leechers), the index must be reprocessed.

I'm aware that certain combinations of filters and orders can be slow, but for most performance is acceptable. For example, ordering a large filtered result by size (ascending) seems very slow, though descending seems okay - I've experimented with a few indexing and database tweaks without much luck and so decided to leave as-is, it can possibly be addressed later or it may be an inherent limitation of Postgres.